plugins/Utils Fixup type of devconfig prop

+++ plugins/Utils/deviceconfig.cpp
@@ -114,17 +114,17 @@ bool DeviceConfig::supportsMultiColorLed() const
     return m_info->contains("supportsMultiColorLed");
 }
 
-quint8 DeviceConfig::sensorLocationX() const
+quint16 DeviceConfig::sensorLocationX() const
 {
     return QString::fromStdString(m_info->get("udfpLocationX", "0")).toUInt();
 }
 
-quint8 DeviceConfig::sensorLocationY() const
+quint16 DeviceConfig::sensorLocationY() const
 {
     return QString::fromStdString(m_info->get("udfpLocationY", "0")).toUInt();
 }
 
-quint8 DeviceConfig::sensorRadius() const
+quint16 DeviceConfig::sensorRadius() const
 {
     return QString::fromStdString(m_info->get("udfpSensorRadius", "0")).toUInt();
 }
